+++ src/main/java/open/commons/spring/web/servlet/method/annotation/DefaultGlobalExceptionHandler.java
@@ -28,6 +28,8 @@ package open.commons.spring.web.servlet.method.annotation;
 
 import javax.validation.ConstraintViolationException;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.AnnotatedElementUtils;
 import org.springframework.core.annotation.Order;
@@ -36,6 +38,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
@@ -46,6 +49,7 @@ import open.commons.spring.web.servlet.BadRequestException;
 import open.commons.spring.web.servlet.InternalServerException;
 import open.commons.spring.web.servlet.NotFoundException;
 import open.commons.spring.web.servlet.UnauthorizedAccessException;
+import open.commons.spring.web.servlet.binder.ExceptionHttpStatusBinder;
 import open.commons.spring.web.utils.WebUtils;
 
 /**
@@ -60,7 +64,11 @@ public class DefaultGlobalExceptionHandler extends ResponseEntityExceptionHandle
 
     public static final TripleFunction<WebRequest, Exception, HttpStatus, FIFOMap<String, Object>> FN_CREATE_ENTITY_DEFAULT = WebUtils::createEntity;
 
+    /** {@link RequestBody} 정보 생성 함수 */
     private final TripleFunction<WebRequest, Exception, HttpStatus, FIFOMap<String, Object>> FN_CREATE_ENTITY;
+
+    /** {@link Throwable}과 {@link HttpStatus} 매핑 정보 */
+    private ExceptionHttpStatusBinder exceptionHttpStatusBinder;
 
     /**
      * <br>
@@ -71,12 +79,34 @@ public class DefaultGlobalExceptionHandler extends ResponseEntityExceptionHandle
      * ------------------------------------------
      * 2020. 1. 17.		박준홍			최초 작성
      * </pre>
-     *
+     * 
      * @since 2020. 1. 17.
      * @version 0.2.3
      */
     public DefaultGlobalExceptionHandler() {
-        this(null);
+        this(null, FN_CREATE_ENTITY_DEFAULT);
+    }
+
+    /***
+     * 
+     * <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2025. 5. 28.		박준홍			최초 작성
+     * </pre>
+     *
+     * @param exceptionHttpStatusBinder
+     *            {@link Throwable}과 {@link HttpStatus} 매핑 정보
+     *
+     * @since 2025. 5. 28.
+     * @version 0.8.0
+     * @author parkjunhong77@gmail.com
+     */
+    public DefaultGlobalExceptionHandler(@Qualifier(ExceptionHttpStatusBinder.BEAN_QUALIFIER) ExceptionHttpStatusBinder exceptionHttpStatusBinder) {
+        this(exceptionHttpStatusBinder, FN_CREATE_ENTITY_DEFAULT);
     }
 
     /**
@@ -86,7 +116,9 @@ public class DefaultGlobalExceptionHandler extends ResponseEntityExceptionHandle
      * ------------------------------------------
      * 2025. 4. 17.     박준홍         최초 작성
      * </pre>
-     *
+     * 
+     * @param exceptionHttpStatusBinder
+     *            {@link Throwable}과 {@link HttpStatus} 매핑 정보
      * @param funcCreateEntity
      *            응답 객체 생성 함수.
      *
@@ -94,8 +126,10 @@ public class DefaultGlobalExceptionHandler extends ResponseEntityExceptionHandle
      * @version 0.8.0
      * @author parkjunhong77@gmail.com
      */
-    public DefaultGlobalExceptionHandler(TripleFunction<WebRequest, Exception, HttpStatus, FIFOMap<String, Object>> funcCreateEntity) {
-        FN_CREATE_ENTITY = funcCreateEntity != null ? funcCreateEntity : FN_CREATE_ENTITY_DEFAULT;
+    public DefaultGlobalExceptionHandler(ExceptionHttpStatusBinder exceptionHttpStatusBinder,
+            TripleFunction<WebRequest, Exception, HttpStatus, FIFOMap<String, Object>> funcCreateEntity) {
+        this.exceptionHttpStatusBinder = exceptionHttpStatusBinder;
+        this.FN_CREATE_ENTITY = funcCreateEntity != null ? funcCreateEntity : FN_CREATE_ENTITY_DEFAULT;
     }
 
     /**
@@ -134,6 +168,7 @@ public class DefaultGlobalExceptionHandler extends ResponseEntityExceptionHandle
      * 2022. 12. 01.    박준홍         {@link NotFoundException} 추가.
      * 2025. 5. 19.     박준홍         {@link UnauthorizedAccessException} 추가.
      * 2025. 5. 28      박준홍         {@link #resolveAnnotatedResponseStatus(Exception, HttpStatus)} 적용
+     * 2025. 5. 28      박준홍         deprecated 처리됨.
      * </pre>
      *
      * @param ex
@@ -143,13 +178,15 @@ public class DefaultGlobalExceptionHandler extends ResponseEntityExceptionHandle
      * @since 2020. 1. 17.
      * @version 0.2.3
      * @author Park_Jun_Hong_(parkjunhong77@gmail.com)
+     * 
+     * @deprecated {@link #handleAllExceptions(Exception, WebRequest)} 일괄적으로 {@link Throwable}를 처리함. 다음 버전에서 삭제될 예정.
      */
-    @ExceptionHandler(value = { //
-            BadRequestException.class, //
-            ConstraintViolationException.class, //
-            NotFoundException.class, //
-            UnauthorizedAccessException.class //
-    })
+    // @ExceptionHandler(value = { //
+    // BadRequestException.class, //
+    // ConstraintViolationException.class, //
+    // NotFoundException.class, //
+    // UnauthorizedAccessException.class //
+    // })
     public ResponseEntity<Object> handle4xxException(Exception ex, WebRequest request) {
 
         HttpStatus status = resolveAnnotatedResponseStatus(ex, HttpStatus.BAD_REQUEST);
@@ -169,9 +206,8 @@ public class DefaultGlobalExceptionHandler extends ResponseEntityExceptionHandle
      * 2020. 7. 30.     박준홍         {@link InternalServerException} 추가
      * 2025. 5. 28      박준홍         {@link UnsupportedOperationException} 처리 분리
      * 2025. 5. 28      박준홍         {@link #resolveAnnotatedResponseStatus(Exception, HttpStatus)} 적용
+     * 2025. 5. 28      박준홍         deprecated 처리됨.
      * </pre>
-     *
-     * 
      * 
      * @param ex
      * @param request
@@ -180,25 +216,70 @@ public class DefaultGlobalExceptionHandler extends ResponseEntityExceptionHandle
      * @since 2020. 1. 17.
      * @version 0.2.3
      * @author Park_Jun_Hong_(parkjunhong77@gmail.com)
+     * 
+     * @deprecated {@link #handleAllExceptions(Exception, WebRequest)} 일괄적으로 {@link Throwable}를 처리함. 다음 버전에서 삭제될 예정.
      */
-    @ExceptionHandler(value = { //
-            NullPointerException.class, //
-            IllegalArgumentException.class, //
-            IllegalStateException.class, //
-            InternalServerException.class, //
-            UnsupportedOperationException.class, //
-            RuntimeException.class, //
-            Exception.class, // eclipse-javadoc:%E2%98%82=open-commons-spring-web/src%5C/main%5C/java%3Copen
-    })
+    // @ExceptionHandler(value = { //
+    // NullPointerException.class, //
+    // IllegalArgumentException.class, //
+    // IllegalStateException.class, //
+    // InternalServerException.class, //
+    // UnsupportedOperationException.class, //
+    // RuntimeException.class, //
+    // Exception.class, // eclipse-javadoc:%E2%98%82=open-commons-spring-web/src%5C/main%5C/java%3Copen
+    // })
     public ResponseEntity<Object> handle5xxException(Exception ex, WebRequest request) {
 
         HttpStatus status = null;
-
         if (UnsupportedOperationException.class.equals(ex.getClass())) {
             status = HttpStatus.SERVICE_UNAVAILABLE;
         } else {
             status = resolveAnnotatedResponseStatus(ex, HttpStatus.INTERNAL_SERVER_ERROR);
         }
+
+        FIFOMap<String, Object> entity = this.FN_CREATE_ENTITY.apply(request, ex, status);
+
+        return handleExceptionInternal(ex, entity, new HttpHeaders(), status, request);
+    }
+
+    /**
+     * 
+     * <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2025. 5. 28.		박준홍			최초 작성
+     * </pre>
+     *
+     * @param ex
+     * @param request
+     * @return
+     *
+     * @since 2025. 5. 28.
+     * @version 0.8.0
+     * @author Park, Jun-Hong parkjunhong77@gmail.com
+     * 
+     * @see #resolveAnnotatedResponseStatus
+     * 
+     */
+    @ExceptionHandler(value = { Throwable.class })
+    public ResponseEntity<Object> handleAllExceptions(Exception ex, WebRequest request) {
+
+        HttpStatus status = resolveAnnotatedResponseStatus(ex, null);
+        if (status == null) {
+            // start - 기존에 CaseByCase로 처리되던 예외클래스 지원. : 2025. 5. 28. 오후 4:46:04
+            // 4xx:
+            if (ConstraintViolationException.class.equals(ex.getClass())) {
+                status = HttpStatus.BAD_REQUEST;
+            }
+            // 5xx
+            else if (UnsupportedOperationException.class.equals(ex.getClass())) {
+                status = HttpStatus.SERVICE_UNAVAILABLE;
+            } //
+        }
+        // end - 기존에 CaseByCase로 처리되던 예외클래스 지원. : 2025. 5. 28. 오후 4:46:04
 
         FIFOMap<String, Object> entity = this.FN_CREATE_ENTITY.apply(request, ex, status);
 
@@ -247,8 +328,39 @@ public class DefaultGlobalExceptionHandler extends ResponseEntityExceptionHandle
      * 
      * @see ResponseStatus
      */
-    protected static HttpStatus resolveAnnotatedResponseStatus(Exception ex, HttpStatus defaultStatus) {
-        ResponseStatus resStatus = AnnotatedElementUtils.findMergedAnnotation(ex.getClass(), ResponseStatus.class);
-        return resStatus != null ? resStatus.code() : defaultStatus;
+    protected HttpStatus resolveAnnotatedResponseStatus(Exception ex, HttpStatus defaultStatus) {
+
+        HttpStatus httpStatus = this.exceptionHttpStatusBinder != null ? this.exceptionHttpStatusBinder.resolveHttpStatus(ex.getClass()) : null;
+        if (httpStatus != null) {
+            return httpStatus;
+        } else {
+            ResponseStatus resStatus = AnnotatedElementUtils.findMergedAnnotation(ex.getClass(), ResponseStatus.class);
+            return resStatus != null ? resStatus.code() : defaultStatus;
+        }
+    }
+
+    /**
+     * <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2025. 5. 28.		박준홍			최초 작성
+     * </pre>
+     *
+     * @param exceptionHttpStatusBinder
+     *            the exceptionHttpStatusBinder to set
+     *
+     * @since 2025. 5. 28.
+     * @version 0.8.0
+     * @author parkjunhong77@gmail.com
+     *
+     * @see #exceptionHttpStatusBinder
+     */
+    @Qualifier(ExceptionHttpStatusBinder.BEAN_QUALIFIER)
+    @Autowired
+    public void setExceptionHttpStatusBinder(ExceptionHttpStatusBinder exceptionHttpStatusBinder) {
+        this.exceptionHttpStatusBinder = exceptionHttpStatusBinder;
     }
 }